feat: add HomeBoard
