+++ src/app/page.tsx
@@ -1,27 +1,9 @@
-import prisma from "../../prisma";
-import MenuBoard from "@/components/menu-board/menu-board.component";
-
-// import TestComponent from "@/components/test.component";
+import HomeBoard from "@/components/home-board/home-board.component";
 
 export default async function Page() {
-  const getQuestions = async () => {
-    const questions = await prisma.question.findMany({
-      include: {
-        additionalQuestion: true,
-        choices: {
-          include: {
-            choice: true,
-          },
-        },
-      },
-    });
-    return {
-      props: { questions },
-    };
-  };
-  const { props } = await getQuestions();
-
-  const params: Question[] = props.questions;
-
-  return <MenuBoard params={params} />;
+  return (
+    <div>
+      <HomeBoard />
+    </div>
+  );
 }
