+++ src/components/home-board/home-board.module.scss
@@ -1,0 +1,11 @@
+.homeBoardContainer {
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+}
+.homeBoard {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
