+++ src/components/home-btn/home-btn.module.scss
@@ -1,0 +1,8 @@
+.homeBtn {
+  height: 7rem;
+  font-size: x-large;
+  background: #ececec;
+  border: 0;
+  padding: 1rem 2rem;
+  cursor: pointer;
+}
